Moved WIT media types and resource into its own file
