+++ design/resources.go
@@ -5,89 +5,6 @@ import (
 	a "github.com/goadesign/goa/design/apidsl"
 )
 
-var _ = a.Resource("workitemtype", func() {
-
-	a.BasePath("/workitemtypes")
-
-	a.Action("show", func() {
-
-		a.Routing(
-			a.GET("/:name"),
-		)
-		a.Description("Retrieve work item type with given name.")
-		a.Params(func() {
-			a.Param("name", d.String, "name")
-		})
-		a.Response(d.OK, func() {
-			a.Media(workItemType)
-		})
-		a.Response(d.NotFound, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-
-	a.Action("create", func() {
-		a.Security("jwt")
-		a.Routing(
-			a.POST(""),
-		)
-		a.Description("Create work item type.")
-		a.Payload(CreateWorkItemTypePayload)
-		a.Response(d.Created, "/workitemtypes/.*", func() {
-			a.Media(workItemType)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.Unauthorized, JSONAPIErrors)
-	})
-
-	a.Action("list", func() {
-		a.Routing(
-			a.GET(""),
-		)
-		a.Description("List work item types.")
-		a.Params(func() {
-			a.Param("page", d.String, "Paging in the format <start>,<limit>")
-		})
-		a.Response(d.OK, func() {
-			a.Media(a.CollectionOf(workItemType))
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-
-	a.Action("list-source-link-types", func() {
-		a.Routing(
-			a.GET("/:name/source-link-types"),
-		)
-		a.Params(func() {
-			a.Param("name", d.String, "name")
-		})
-		a.Description(`Retrieve work item link types where the
-given work item type can be used in the source of the link.`)
-		a.Response(d.OK, func() {
-			a.Media(workItemLinkTypeList)
-		})
-		a.Response(d.NotFound, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-
-	a.Action("list-target-link-types", func() {
-		a.Routing(
-			a.GET("/:name/target-link-types"),
-		)
-		a.Params(func() {
-			a.Param("name", d.String, "name")
-		})
-		a.Description(`Retrieve work item link types where the
-given work item type can be used in the target of the link.`)
-		a.Response(d.OK, func() {
-			a.Media(workItemLinkTypeList)
-		})
-		a.Response(d.NotFound, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-})
-
 var _ = a.Resource("status", func() {
 
 	a.DefaultMedia(ALMStatus)
